Allows working copy of Plone Site
The container for the Plone Site working copy is Plone Site itself.

To avoid permission errors, the Plone Site working copy is created as
the Document type. But right after the Document is created, its
portal_type is changed to "Plone Site", so that the schema of the
working copy is the same as the Plone Site schema.

- Prevents Plone Site permissions from being changed on checkin.